+++ ui/src/shared/components/TimeSeries.tsx
@@ -32,6 +32,7 @@ import {
   isDemoDataAvailabilityError,
   demoDataError,
 } from 'src/cloud/utils/demoDataErrors'
+import {hashCode} from 'src/queryCache/actions'
 
 // Constants
 import {
@@ -43,6 +44,7 @@ import {TIME_RANGE_START, TIME_RANGE_STOP} from 'src/variables/constants'
 
 // Actions
 import {notify as notifyAction} from 'src/shared/actions/notifications'
+import {setQueryResultsByQueryID} from 'src/queryCache/actions'
 
 // Types
 import {
@@ -87,6 +89,7 @@ interface OwnProps {
 
 interface DispatchProps {
   notify: typeof notifyAction
+  onSetQueryResultsByQueryID: typeof setQueryResultsByQueryID
 }
 
 type Props = StateProps & OwnProps & DispatchProps
@@ -181,7 +184,13 @@ class TimeSeries extends Component<Props & WithRouterProps, State> {
   }
 
   private reload = async () => {
-    const {variables, notify, check, buckets} = this.props
+    const {
+      buckets,
+      check,
+      notify,
+      onSetQueryResultsByQueryID,
+      variables,
+    } = this.props
     const queries = this.props.queries.filter(({text}) => !!text.trim())
 
     if (!queries.length) {
@@ -270,6 +279,11 @@ class TimeSeries extends Component<Props & WithRouterProps, State> {
       }
 
       this.pendingReload = false
+      const queryText = queries.map(({text}) => text).join('')
+      const queryID = hashCode(queryText)
+      if (queryID && files.length) {
+        onSetQueryResultsByQueryID(queryID, files)
+      }
 
       this.setState({
         giraffeResult,
@@ -346,6 +360,7 @@ const mstp = (state: AppState, props: OwnProps): StateProps => {
 
 const mdtp: DispatchProps = {
   notify: notifyAction,
+  onSetQueryResultsByQueryID: setQueryResultsByQueryID,
 }
 
 export default connect<StateProps, {}, OwnProps>(